types: ReadableToSyncIO maps ArrayBufferView to Uint8Array
Update the JSDoc for stdio/stdout/stderr to note the ArrayBufferView sink
is spawnSync-only, add the ArrayBufferView arm to ReadableToSyncIO so
SyncSubprocess.stdout/stderr are typed as Uint8Array instead of
undefined, drop the now-unneeded cast in spawn.test.ts, and add a
type-level fixture check.

+++ test/js/bun/spawn/spawn.test.ts
@@ -279,7 +279,7 @@ for (let [gcTick, label] of [
         const text2 = new TextDecoder().decode(stdout_buffer);
         expect(text).toBe("hello world\n");
         expect(text2).toBe("hello world\n\u0000\u0000\u0000");
-        expect((stdout as Uint8Array).buffer).toBe(stdout_buffer.buffer);
+        expect(stdout.buffer).toBe(stdout_buffer.buffer);
         gcTick();
       });
